refactor: move softball-only components to components/softball/
Moved 10 softball-specific components from components/sports/ into
components/softball/. Only auth-button remains in components/sports/
since it's shared with the generic sport page.
