+++ app/softball/page.tsx
@@ -3,10 +3,10 @@ import { createClient } from "@/lib/supabase/server";
 import { getUser } from "@/lib/supabase/user";
 import { Tabs, TabsList, TabsTrigger, TabsContent } from "@/components/ui/tabs";
 import { Settings } from "lucide-react";
-import SessionCard from "@/components/sports/session-card";
-import TeamAccessBanner from "@/components/sports/team-access-banner";
-import SignInPrompt from "@/components/sports/sign-in-prompt";
-import SoftballPageShell from "@/components/sports/softball-page-shell";
+import SessionCard from "@/components/softball/session-card";
+import TeamAccessBanner from "@/components/softball/team-access-banner";
+import SignInPrompt from "@/components/softball/sign-in-prompt";
+import SoftballPageShell from "@/components/softball/softball-page-shell";
 import { Button } from "@/components/ui/button";
 import { sportsConfig } from "@/lib/sports-config";
 import { createAdminClient } from "@/lib/supabase/admin";
