+++ components/softball/session-card.tsx
